Made implicit references explicit, fixed many test cases, reduced dependencies
UMLchange profile seems to be broken

git-svn-id: https://svn.uni-koblenz.de/rgse/carisma/trunk@9392 b99f87ec-93df-11e5-8f70-1b308f1eeebc

+++ projects/carisma/trunk/tests/carisma.umlchange.efficiency.tests/src/carisma/check/umlchange/efficiency/TestTest.java
@@ -1,13 +1,13 @@
 package carisma.check.umlchange.efficiency;
 
-import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.uml2.uml.Model;
 
 public class TestTest {
 	
+	//TODO: Has to be executed as plugin, convert to JUnit Plugin Test?
 	public static void main(String[] args) {
 		LockedStatusModelCreater mc = new LockedStatusModelCreater();
-		Model model = (Model) mc.getNewModel("TestModel", 100, (float) 10).getContents().get(0);
+		Model model = (Model) mc.getNewModel("TestModel", 100, 10).getContents().get(0);
 		System.out.println("Success");
 		
 	}
